+++ Twilio/Rest/Api/V2010/Account/Sip/Domain/AuthTypes/AuthTypeCallsList.php
@@ -29,7 +29,7 @@ class AuthTypeCallsList extends ListResource {
      * Construct the AuthTypeCallsList
      * 
      * @param Version $version Version that contains the resource
-     * @param string $accountSid The unique id of the account that sent the message
+     * @param string $accountSid The unique id of the account that sent the call
      * @param string $domainSid A string that uniquely identifies the SIP Domain
      * @return \Twilio\Rest\Api\V2010\Account\Sip\Domain\AuthTypes\AuthTypeCallsList 
      */