+++ Twilio/Rest/Pricing/V2/Voice/CountryList.php
@@ -7,60 +7,30 @@
  * /       /
  */
 
-namespace Twilio\Rest\Api\V2010\Account\Sip\Domain\AuthTypes\AuthTypeCalls;
+namespace Twilio\Rest\Pricing\V2\Voice;
 
 use Twilio\ListResource;
 use Twilio\Values;
 use Twilio\Version;
 
-class AuthCallsCredentialListMappingList extends ListResource {
+class CountryList extends ListResource {
     /**
-     * Construct the AuthCallsCredentialListMappingList
+     * Construct the CountryList
      * 
      * @param Version $version Version that contains the resource
-     * @param string $accountSid The unique id of the account that sent the call
-     * @param string $domainSid A string that uniquely identifies the SIP Domain
-     * @return \Twilio\Rest\Api\V2010\Account\Sip\Domain\AuthTypes\AuthTypeCalls\AuthCallsCredentialListMappingList 
+     * @return \Twilio\Rest\Pricing\V2\Voice\CountryList 
      */
-    public function __construct(Version $version, $accountSid, $domainSid) {
+    public function __construct(Version $version) {
         parent::__construct($version);
 
         // Path Solution
-        $this->solution = array('accountSid' => $accountSid, 'domainSid' => $domainSid, );
+        $this->solution = array();
 
-        $this->uri = '/Accounts/' . rawurlencode($accountSid) . '/SIP/Domains/' . rawurlencode($domainSid) . '/Auth/Calls/CredentialListMappings.json';
+        $this->uri = '/Voice/Countries';
     }
 
     /**
-     * Create a new AuthCallsCredentialListMappingInstance
-     * 
-     * @param string $credentialListSid A string that uniquely identifies this
-     *                                  credential list resource
-     * @return AuthCallsCredentialListMappingInstance Newly created
-     *                                                AuthCallsCredentialListMappingInstance
-     * @throws TwilioException When an HTTP error occurs.
-     */
-    public function create($credentialListSid) {
-        $data = Values::of(array('CredentialListSid' => $credentialListSid, ));
-
-        $payload = $this->version->create(
-            'POST',
-            $this->uri,
-            array(),
-            $data
-        );
-
-        return new AuthCallsCredentialListMappingInstance(
-            $this->version,
-            $payload,
-            $this->solution['accountSid'],
-            $this->solution['domainSid']
-        );
-    }
-
-    /**
-     * Streams AuthCallsCredentialListMappingInstance records from the API as a
-     * generator stream.
+     * Streams CountryInstance records from the API as a generator stream.
      * This operation lazily loads records as efficiently as possible until the
      * limit
      * is reached.
@@ -86,7 +56,7 @@ class AuthCallsCredentialListMappingList extends ListResource {
     }
 
     /**
-     * Reads AuthCallsCredentialListMappingInstance records from the API as a list.
+     * Reads CountryInstance records from the API as a list.
      * Unlike stream(), this operation is eager and will load `limit` records into
      * memory before returning.
      * 
@@ -98,21 +68,20 @@ class AuthCallsCredentialListMappingList extends ListResource {
      *                        page_size is defined but a limit is defined, read()
      *                        will attempt to read the limit with the most
      *                        efficient page size, i.e. min(limit, 1000)
-     * @return AuthCallsCredentialListMappingInstance[] Array of results
+     * @return CountryInstance[] Array of results
      */
     public function read($limit = null, $pageSize = null) {
         return iterator_to_array($this->stream($limit, $pageSize), false);
     }
 
     /**
-     * Retrieve a single page of AuthCallsCredentialListMappingInstance records
-     * from the API.
+     * Retrieve a single page of CountryInstance records from the API.
      * Request is executed immediately
      * 
      * @param mixed $pageSize Number of records to return, defaults to 50
      * @param string $pageToken PageToken provided by the API
      * @param mixed $pageNumber Page Number, this value is simply for client state
-     * @return \Twilio\Page Page of AuthCallsCredentialListMappingInstance
+     * @return \Twilio\Page Page of CountryInstance
      */
     public function page($pageSize = Values::NONE, $pageToken = Values::NONE, $pageNumber = Values::NONE) {
         $params = Values::of(array(
@@ -127,16 +96,15 @@ class AuthCallsCredentialListMappingList extends ListResource {
             $params
         );
 
-        return new AuthCallsCredentialListMappingPage($this->version, $response, $this->solution);
+        return new CountryPage($this->version, $response, $this->solution);
     }
 
     /**
-     * Retrieve a specific page of AuthCallsCredentialListMappingInstance records
-     * from the API.
+     * Retrieve a specific page of CountryInstance records from the API.
      * Request is executed immediately
      * 
      * @param string $targetUrl API-generated URL for the requested results page
-     * @return \Twilio\Page Page of AuthCallsCredentialListMappingInstance
+     * @return \Twilio\Page Page of CountryInstance
      */
     public function getPage($targetUrl) {
         $response = $this->version->getDomain()->getClient()->request(
@@ -144,22 +112,17 @@ class AuthCallsCredentialListMappingList extends ListResource {
             $targetUrl
         );
 
-        return new AuthCallsCredentialListMappingPage($this->version, $response, $this->solution);
+        return new CountryPage($this->version, $response, $this->solution);
     }
 
     /**
-     * Constructs a AuthCallsCredentialListMappingContext
+     * Constructs a CountryContext
      * 
-     * @param string $sid Fetch by unique credential list Sid
-     * @return \Twilio\Rest\Api\V2010\Account\Sip\Domain\AuthTypes\AuthTypeCalls\AuthCallsCredentialListMappingContext 
+     * @param string $isoCountry Fetches voice prices for country
+     * @return \Twilio\Rest\Pricing\V2\Voice\CountryContext 
      */
-    public function getContext($sid) {
-        return new AuthCallsCredentialListMappingContext(
-            $this->version,
-            $this->solution['accountSid'],
-            $this->solution['domainSid'],
-            $sid
-        );
+    public function getContext($isoCountry) {
+        return new CountryContext($this->version, $isoCountry);
     }
 
     /**
@@ -168,6 +131,6 @@ class AuthCallsCredentialListMappingList extends ListResource {
      * @return string Machine friendly representation
      */
     public function __toString() {
-        return '[Twilio.Api.V2010.AuthCallsCredentialListMappingList]';
+        return '[Twilio.Pricing.V2.CountryList]';
     }
 }